Walker2014Sampling

Documentation for Walker2014Sampling.

Walker2014Sampling.Walker2014Sampler

Base.rand

Walker2014Sampling.Walker2014Sampler — Type

Walker2014Sampler(k::Int)

The sampler proposed by Walker (2014), with parameter k.

Base.rand — Method

rand(rng::AbstractRNG, s::Walker2014Sampler, p::Function, x::Int)

Draw the next value in a MCMC with target distribution p and current value x, using the sampler s. Both rng and s are modified in the process.